+++ app/project-listing/CHANGELOG.md
@@ -4,7 +4,6 @@
 
 ### Added
 
-- Added 2 projects: random-id, text-case
+- Added 3 projects: random-id, secret-message, text-case
 - Added home page to show projects
-- Added app description
 - Resolved url by environment